pyproject.toml
[tool.poetry]
name = "amb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme="README.rst"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
requests = "^2.23.0"
nose2 = "^0.9.2"
"zope.interface" = "^5.1.0"
pillow = "^7.1.2"
docutils = "^0.16"
pygments = "^2.6.1"
pypiwin32 = "^223"
"kivy_deps.glew" = "^0.2.0"
"kivy_deps.sdl2" = "^0.2.0"
"kivy_deps.gstreamer" = "^0.2.0"
kivy = "^1.11.1"
sqlalchemy = "^1.3.17"
pydub = "^0.24.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sphinx = "^3.0.3"
flake8 = "^3.8.2"
black = "^19.10b0"
nose2 = "^0.9.2"
sphinx_rtd_theme = "^0.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
